Move processing logic into network service
- Refactored the network logic to avoid deep if-nesting
- Added result type for invalid responses
